fixed date format to account for same '%y' output
By displaying the year as a whole rather than the last two digits, screenshots are
differentiated even if the last two digits of the year are the same.
For example, if two screenshots were taken at the exact same time except
for the year, then the file name will be the same without this commit.

Example with hypothetical screenshots taken at different years.
with this commit:
pic-full-2022-12-13_19-35-41.png
pic-full-2522-12-13_19-35-41.png

without this commit:
pic-full-22-12-13_19-35-41.png
pic-full-22-12-13_19-35-41.png