added option to override using awesomewm notifs
useLibnotify = true
With this, even if the current window manager being used is awesomewm,
it will make timer use libnotify (notify-send) to send the notification.

Fixed a bug where the time that the timer went off is not printed after
sending a notification through awesomewm. With this commit, the time
when the timer went off is printed after sending a notification from
awesomewm.