made awesomewm forecast widget work
sb-forecast has a LOCATION variable assigned.
With this, one can modify this variable to be their location
to get the weather for it.

The awesomewm forecast widget will run sb-forecast after clicking it.
sb-forecast will update the weather report as needed. After that,
the widget will open the weather report.