changed mpv thumbnailer scripts
The older thumbnailer scripts was using mpv features that will be
deprecated. Despite this, the scripts seem to not be maintained anymore
(last commit was around 5 years ago).

Replaced the old thumbnailer script with another project that does the
same job but seems maintained (last commit was around 3 months ago).