optionally hibernate when battery is very low
With this commit, when the battery is at danger level,
a menu will appear to choose from hibernating the system or not.

This way, the system won't hibernate by itself without input from the
user. For example, the user may be close to a charger and can quickly
plugin the laptop without being interrupted by the system hibernating.