make triggerTimestamp handle a string with white spaces as 1 argument
Since the argument passed in was not encased with double quotes (""),
an argument with white spaces gets expanded as multiple arguments, each
having different references.

For example, the argument passed into timer
on the command line like this:
timer '1h 40m'...
will result in '1h 40m' being passed into triggerTimestamp as
'1h' and '40m'
$1 refers to '1h' and $2 refers to '40m'. The grep line only looks at
$1. Thus, '40m' is not accessed when setting the variable, minutes.

With this commit, there will only be one argument for triggerTimestamp.
So, with the previous example, '1h 40m' will actually be passed into
triggerTimestamp as '1h 40m' and accessible with $1.